nutr-backend/requirements.txt
Django>=6.0
djangorestframework>=3.15
django-cors-headers>=4.0
firebase-admin>=6.0
mysqlclient>=2.2
celery>=5.3
redis>=5.0
python-decouple>=3.8
django-filter>=24.0
requests>=2.31
